+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added integration with Pop, Livestorm, Chorus & Gong
 - Fix readability of the statusbar text in multi-screen setups (#354)
 - Detect hidden menubar icon (#429)
+- Added feature to snooze the notification
 
 ## Version 3.10.0